Write tests to check keystrokes during studying.
This is in preparation for adding options to edit and delete each
flashcard.

It wasn't immediately obvious how to enter the inputs - seemed like they
were being ignored. However, this was due to trying to enter them like
click.prompt inputs - separated by newlines. For click.pause, as only a
single keystroke is needed to continue, use just a string, one character
for each stroke. Note that the study helper (as currently written) only
checks the second one, after displaying the answer to the question, and
so effectively each input to the test is evaluated at that point.