Combine the main study command and its helper
This was the only main command that had its core logic separated
out into a special helper command. That did not really add anything,
so I just added the helper to the main command, and did the same to
its tests, and deleted the file for the helper and the test file for
it.